+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
+## Unreleased
+### Changed
+- YLP-203 Change menu item "System settings" to "DBL settings"
+- YLP-205 Change menu item "print"
+
 ## 1.5.0 - 2020-10-05  
 ### Changed
 - YLP-152 Design update consent screen.
@@ -335,4 +340,3 @@ It is based on Tidepool Blip 1.27.
    - user automatically disconnected [PT-69]
    - Fix race condition [PT-17]
    - Fix Unknown DOM property class [PT-163]
-